% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hia_data.R
\docType{data}
\name{hia_data}
\alias{hia_data}
\title{Datos de ejemplo para calcular casos atribuíbles a la exposición a contaminantes del aire}
\format{
\subsection{\code{hia_data}}{

A data frame with 16 rows and 5 columns:
\describe{
\item{cvegeo}{Clave del área geográfica}
\item{area_name}{nombre del área geográfica}
\item{population}{Población de 30 años y más del área de estudio}
\item{concentration}{Concentración del contaminante en el área de estudio}
\item{cases}{Total de muertes en población de 30 años y más en el área de estudio}
...
}
}
}
\source{
\url{https://www.who.int/health-topics/health-impact-assessment}
}
\usage{
hia_data
}
\description{
Este conjunto de datos son de carñacter ficticio y unicamente se comparten
para facilitar la comprensión de los comandos y replicar ejemplos ...
}
\keyword{datasets}
