% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hia_no2.R
\name{hia_no2}
\alias{hia_no2}
\title{Esta función realiza la estimación de casos atribuíbles a la exposición a NO2}
\usage{
hia_no2(x, uc = 10, cf, fcr.low = 1.01, fcr.mean = 1.02, fcr.high = 1.04)
}
\arguments{
\item{x}{data frame o tibble.}

\item{uc}{Unidad de Cambio de la FCR en ug/m3. Por defecto el valor es 10.}

\item{cf}{valor contrafactual, generalmente el valor guía de la OMS o NOM.}

\item{fcr.low}{Valor Inferior de la FCR. Por defecto el valor es 1.01.}

\item{fcr.mean}{Valor Promedio de la FCR. Por defecto el valor es 1.02.}

\item{fcr.high}{Valor Superior de la FCR. Por defecto el valor es 1.04.}
}
\value{
Devuelve una tabla con los resultados.
}
\description{
Esta función realiza la estimación de casos atribuíbles a la exposición a NO2
}
\examples{
data(hia_data)

## Usando valores por default
hia_no2(hia_data, cf = 10)

## Especificando una distinta FCR
hia_no2(hia_data, uc = 10, cf = 12, fcr.low = 1.040, fcr.mean = 1.062, fcr.high = 1.083)

}
