% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hia_ozono.R
\name{hia_ozono}
\alias{hia_ozono}
\title{Esta función realiza la estimación de casos atribuíbles a la exposición a Ozono}
\usage{
hia_ozono(x, uc = 10, cf, fcr.low = 1, fcr.mean = 1.01, fcr.high = 1.02)
}
\arguments{
\item{x}{data frame o tibble.}

\item{uc}{Unidad de Cambio de la FCR en ug/m3. Por defecto el valor es 10.}

\item{cf}{valor contrafactual, generalmente el valor guía de la OMS o NOM.}

\item{fcr.low}{Valor Inferior de la FCR. Por defecto el valor es 1.00.}

\item{fcr.mean}{Valor Promedio de la FCR. Por defecto el valor es 1.01.}

\item{fcr.high}{Valor Superior de la FCR. Por defecto el valor es 1.02.}
}
\value{
Devuelve una tabla con los resultados.
}
\description{
Esta función realiza la estimación de casos atribuíbles a la exposición a Ozono
}
\examples{
#' data(hia_data)

## Usando valores por default
hia_ozono(hia_data, cf = 10)

#' ## Especificando una distinta FCR
hia_ozono(hia_data, uc = 10, cf = 12, fcr.low = 1.005, fcr.mean = 1.014, fcr.high = 1.024)

}
