% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hia_pm10.R
\name{hia_pm10}
\alias{hia_pm10}
\title{Esta función realiza la estimación de casos atribuíbles a la exposición a PM10}
\usage{
hia_pm10(x, uc = 10, cf, fcr.low = 1.03, fcr.mean = 1.04, fcr.high = 1.06)
}
\arguments{
\item{x}{data frame o tibble.}

\item{uc}{Unidad de Cambio de la FCR en ug/m3. Por defecto el valor es 10.}

\item{cf}{valor contrafactual, generalmente el valor guía de la OMS o NOM.}

\item{fcr.low}{Valor Inferior de la FCR. Por defecto el valor es 1.03.}

\item{fcr.mean}{Valor Promedio de la FCR. Por defecto el valor es 1.04.}

\item{fcr.high}{Valor Superior de la FCR. Por defecto el valor es 1.06.}
}
\value{
Devuelve una tabla con los resultados.
}
\description{
Esta función realiza la estimación de casos atribuíbles a la exposición a PM10
}
\examples{
data(hia_data)

## Usando valores por default
hia_pm10(hia_data, cf = 15)

## Especificando una distinta FCR
hia_pm10(hia_data, uc = 10, cf = 15, fcr.low = 1.040, fcr.mean = 1.062, fcr.high = 1.083)

}
