% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hia_pm25.R
\name{hia_pm25}
\alias{hia_pm25}
\title{Esta función realiza la estimación de casos atribuíbles a la exposición a
contaminantes del aire}
\usage{
hia_pm25(x, uc = 10, cf, fcr.low = 1.06, fcr.mean = 1.08, fcr.high = 1.09)
}
\arguments{
\item{x}{data frame o tibble.}

\item{uc}{Unidad de Cambio de la FCR en ug/m3. Por defecto el valor es 10.}

\item{cf}{valor contrafactual, generalmente el valor guía de la OMS o NOM.}

\item{fcr.low}{Valor Inferior de la FCR. Por defecto el valor es 1.06.}

\item{fcr.mean}{Valor Promedio de la FCR. Por defecto el valor es 1.08.}

\item{fcr.high}{Valor Superior de la FCR. Por defecto el valor es 1.09.}
}
\value{
Devuelve una tabla con los resultados
}
\description{
Esta función realiza la estimación de casos atribuíbles a la exposición a
contaminantes del aire
}
\examples{
mydata <-
  data.frame(cvegeo = c("01", "02", "03", "04", "05"),
             area_name = c("area_1", "area_2", "area_3", "area_4", "area_5"),
             population = c(749982, 400161, 417416, 60847, 199224),
             concentration = c(21.8, 25.8, 24.5, 21.6, 17.7),
             cases = c(4070, 3004, 2959, 4038, 806))

## Usando valores por default
hia_pm25(mydata, cf = 10)

## Especificando una distinta FCR
hia_pm25(mydata, uc = 10, cf = 12, fcr.low = 1.040, fcr.mean = 1.062, fcr.high = 1.083)

}
