% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hia_so2.R
\name{hia_so2}
\alias{hia_so2}
\title{Esta función realiza la estimación de casos atribuíbles a la exposición a SO2}
\usage{
hia_so2(x, uc = 10, cf, fcr.low = 1.0046, fcr.mean = 1.0059, fcr.high = 1.0071)
}
\arguments{
\item{x}{data frame o tibble.}

\item{uc}{Unidad de Cambio de la FCR en ug/m3. Por defecto el valor es 10.}

\item{cf}{valor contrafactual, generalmente el valor guía de la OMS o NOM.}

\item{fcr.low}{Valor Inferior de la FCR. Por defecto el valor es 1.0046.}

\item{fcr.mean}{Valor Promedio de la FCR. Por defecto el valor es 1.0059.}

\item{fcr.high}{Valor Superior de la FCR. Por defecto el valor es 1.0071.}
}
\value{
Devuelve una tabla con los resultados.
}
\description{
Esta función realiza la estimación de casos atribuíbles a la exposición a SO2
}
\examples{
data(hia_data)

 ## Usando valores por default
 hia_so2(hia_data, cf = 10)

 ## Especificando una distinta FCR
 #' hia_no2(hia_data, uc = 10, cf = 12, fcr.low = 1.040, fcr.mean = 1.062, fcr.high = 1.083)

}
